+++ PotteryBarn/ComfyLib/Extensions/CodeMatcherExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace ComfyLib;
+namespace ComfyLib;
 
 using System.Collections.Generic;
 using System.Reflection.Emit;
@@ -7,8 +7,9 @@ using HarmonyLib;
 
 public static class CodeMatcherExtensions {
   public static CodeMatcher ExtractLabels(this CodeMatcher matcher, out List<Label> labels) {
-    labels = new(matcher.Labels);
+    labels = [.. matcher.Labels];
     matcher.Labels.Clear();
+
     return matcher;
   }
 }