+++ PotteryBarn/ComfyLib/Extensions/ConfigFileExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace ComfyLib;
+namespace ComfyLib;
 
 using System;
 using System.Collections.Generic;