+++ PotteryBarn/ComfyLib/Extensions/UnityExtensions.cs
@@ -1,11 +1,11 @@
-﻿namespace ComfyLib;
+namespace ComfyLib;
 
 using System;
 
 public static class ChatExtensions {
   public static void AddMessage(this Chat chat, object obj) {
     if (chat) {
-      chat.AddString($"{obj}");
+      chat.AddString(obj.ToString());
       chat.m_hideTimer = 0f;
     }
   }