+++ PotteryBarn/ComfyLib/Utils/IsExternalInit.cs
@@ -1,4 +1,4 @@
-﻿namespace System.Runtime.CompilerServices;
+namespace System.Runtime.CompilerServices;
 
 using System.ComponentModel;