+++ PotteryBarn/Config/PluginConfig.cs
@@ -1,4 +1,4 @@
-﻿namespace PotteryBarn;
+namespace PotteryBarn;
 
 using BepInEx.Configuration;