+++ PotteryBarn/Core/Utils/CustomFieldUtils.cs
@@ -1,16 +1,16 @@
-﻿namespace PotteryBarn;
+namespace PotteryBarn;
 
 public static class CustomFieldUtils {
-  public static readonly int HasFieldsHashCode = "HasFields".GetStableHashCode();
-  public static readonly int HasFieldsDestructibleHashCode = "HasFieldsDestructible".GetStableHashCode();
-  public static readonly int DestructibleSpawnWhenDestroyedHashCode =
-      "Destructible.m_spawnWhenDestroyed".GetStableHashCode();
+  public const int HasFieldsHash = -310439593;                        // HasFields
+  public const int HasFieldsDestructibleHash = 949312639;             // HasFieldsDestructible
+  public const int DestructibleSpawnWhenDestroyedHash = -1786945712;  // Destructible.m_spawnWhenDestroyed
 
   public static void SetDestructibleFields(Destructible destructible, string spawnWhenDestroyed) {
     ZDO zdo = destructible.m_nview.m_zdo;
-    zdo.Set(HasFieldsHashCode, true);
-    zdo.Set(HasFieldsDestructibleHashCode, true);
-    zdo.Set(DestructibleSpawnWhenDestroyedHashCode, spawnWhenDestroyed);
+
+    zdo.Set(HasFieldsHash, true);
+    zdo.Set(HasFieldsDestructibleHash, true);
+    zdo.Set(DestructibleSpawnWhenDestroyedHash, spawnWhenDestroyed);
 
     destructible.m_nview.LoadFields();
   }