+++ PotteryBarn/Patches/PiecePatch.cs
@@ -1,4 +1,4 @@
-﻿namespace PotteryBarn;
+namespace PotteryBarn;
 
 using HarmonyLib;