+++ PotteryBarn/PotteryBarn.cs
@@ -1,4 +1,4 @@
-﻿namespace PotteryBarn;
+namespace PotteryBarn;
 
 using System.Reflection;
 
@@ -15,7 +15,7 @@ using static PluginConfig;
 public sealed class PotteryBarn : BaseUnityPlugin {
   public const string PluginGuid = "redseiko.valheim.potterybarn";
   public const string PluginName = "PotteryBarn";
-  public const string PluginVersion = "1.20.0";
+  public const string PluginVersion = "1.21.0";
 
   void Awake() {
     BindConfig(Config);